llm-demo/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Optional AI/ML dependencies (uncomment when needed)
# torch==2.1.0
# torchvision==0.16.0  
# diffusers==0.24.0
# transformers==4.36.0
# opencv-python==4.8.1.78
# numpy==1.24.3